pyproject.toml
[project]
name = "py-faiss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "chardet>=5.2.0",
    "datetime>=5.5",
    "faiss-cpu>=1.11.0.post1",
    "fastapi[standard]>=0.116.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pathlib>=1.0.1",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.35.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[project.scripts]
py-faiss = "py_faiss.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]
